NavChanges
Primarily page name changes on the Client App.

NavMenu - changed to GiftAll and renamed links/icons

CreateHome - previously Home

Homepage - basic homepage outline w/ bottom nav

Invite - invite user outline

LoginGiftAll - login page
