+++ ClientApp/ClientApp/Components/Layout/NavMenu.razor.css
@@ -69,12 +69,7 @@
         align-items: center;
         line-height: 3rem;
         font-style: italic;
-    }
-
-        .nav-item ::deep a.active {
-            background-color: rgba(255,255,255,0.37);
-            color: white;
-        }
+    } 
 
         .nav-item ::deep a:hover {
             background-color: rgba(255,255,255,0.1);
