+++ ClientApp/ClientApp/Components/Pages/Homepage.razor
@@ -1,0 +1,12 @@
+@page "/homepage"
+
+<div class="homepageGiftAll"> 
+
+    <h1 class="createTooltip"> Click the g button below to create a house! </h1> <!-- hidden after #houses>1 -->
+
+    <div class="bottomNav">
+        <div class="profileButton"> <a class="profile"> </a> </div>
+        <div class="plusCreateHouse"> <a class="create"> <img src="/Images/gift.png" asp-append-version="true" width="50" height="50" /> </a> </div>
+    </div> <!-- end bottom nav-->
+
+</div> <!-- end homepageGiftAll -->
