+++ ClientApp/ClientApp/Components/Pages/Homepage.razor.css
@@ -1,0 +1,51 @@
+* {
+    margin: 0 auto;
+    text-align: center;
+    font-family: 'Courier New';
+    border-radius: 12px;
+}
+.homepageGiftAll {
+    background-color: pink;
+    position: relative;
+    width: 100%;
+    height: 630px;
+}
+
+.bottomNav {
+    overflow: auto;
+    background-color: coral;
+    height:75px;
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    opacity: 50%;
+}
+
+.profileButton {
+    background-color: mistyrose;
+    clip-path: 50%;
+    width: 50px;
+    height: 50px;
+    float: right;
+    margin-top: 15px;
+    margin-right: 5px;
+}
+
+.plusCreateHouse {
+    background-color: aquamarine;
+    clip-path: 50%;
+    width: 50px;
+    height: 50px;
+    margin-top: 15px;
+}
+
+.create {
+    background-image:url(/Images/gift.png);
+}
+
+.createTooltip {
+    padding: 10px;
+    position: absolute;
+    top: 25%;
+    width: 100%;
+}
