+++ ClientApp/ClientApp/Components/Pages/Invite.razor
@@ -1,0 +1,27 @@
+@page "/invite"
+
+<div class="GiftInvite">
+
+    <div class="inviteBubble">
+        <h1> Invite to ~GiftHouseName</h1>
+        <form>
+            <label for="email"> Enter Recipient's Email: </label>
+            <input type="text" id="recipientsEmail" name="email" required><br>
+            <br>
+
+            <div class="emailList"> 
+                <p> email@email.com </p>
+                <p> email@email.com </p>
+            </div> <!-- end emailList -->
+
+            <input type="submit" value="Click to Invite!">
+        </form>
+
+        <!-- <button onclick="Invite" class="inviteRecipient"> Click to Invite! </button> -->
+        <div class="confirmationStatus">
+            <p class="inviteStatus"> ~PendingInvite </p>
+        </div> <!-- end confirmationStatus -->
+
+    </div> <!-- end inviteBubble -->
+
+</div> <!-- end GiftInvite -->
