+++ ClientApp/ClientApp/Components/Pages/Invite.razor.css
@@ -1,0 +1,63 @@
+* {
+    margin: 0 auto;
+    box-sizing: border-box;
+    text-align: center;
+    border-radius: 12px;
+}
+
+.GiftInvite {
+    margin-top: 2%;
+    width: 100%;
+    line-height: 1.8;
+}
+
+h1 {
+    margin-top: 2px;
+    width: 90%;
+}
+
+.inviteBubble {
+    padding-top: 1%;
+    width: 75%;
+    border-radius: 15px 50px 50px;
+    background-color: lightpink;
+}
+
+label, input {
+    font-size: 32px;
+}
+
+    button, input[type=submit] {
+        background-color: plum;
+        border: none;
+        padding: 5px;
+    }
+
+.confirmationStatus {
+    visibility: hidden;
+}
+
+i.material-icons {
+    font-size: 14px;
+}
+
+input[type=text] {
+    width: 85%;
+    border: 2px solid pink;
+}
+
+    input[type=text]:focus {
+        background-color: lightgrey;
+    }
+
+input[type=submit] {
+    margin-top: 2px;
+}
+
+/* Responsive layout - when the screen is less than 600px wide, make the two columns stack on top of each other instead of next to each other */
+@media screen and (max-width: 600px) {
+    input[type=submit] {
+        width: 100%;
+        margin-top: 5px;
+    }
+}
