+++ ClientApp/ClientApp/Components/Pages/LoginGiftAll.razor
@@ -1,0 +1,8 @@
+@page "/"
+
+<div class="GiftAllLogin">
+    <div class="titlePage"> <h1> GIFT ALL </h1>
+
+    <button onclick="@Login" class="giftLogin"> LOGIN </button> </div>
+
+</div> <!-- end GiftAllLogin -->
