+++ ClientApp/ClientApp/Components/Pages/LoginGiftAll.razor.cs
@@ -5,31 +5,14 @@ using Microsoft.AspNetCore.Components.Forms;
 
 namespace ClientApp.Components.Pages
 {
-    public partial class Home
+    public partial class LoginGiftAll
     {
-        [Inject] LoginService LoginService { get; set; } = null!;
-
-        public GiftHouse? GiftHouseData { get; set; }
-        public EditContext? EditContext { get; set; }
-        public string MessageStore { get; set; } = "";
+        [Inject] LoginService LoginService { get; set; } = null!; 
 
         private bool _isLoggingIn = false;
-
-        protected override void OnInitialized()
-        {
-            GiftHouseData = new GiftHouse();
-            EditContext = new EditContext(GiftHouseData);
-        }
-
-        public void Submit()
-        {
-            MessageStore = $"{GiftHouseData?.CreatorName} created House {GiftHouseData?.HouseName}";
-            GiftHouseData = new GiftHouse();
-        }
-
         public async void Login()
         {
-            if (_isLoggingIn) 
+            if (_isLoggingIn)
             {
                 return;
             }
