+++ ClientApp/ClientApp/Components/Pages/LoginGiftAll.razor.css
@@ -1,0 +1,34 @@
+﻿* {
+    margin: 0 auto;
+    text-align: center;
+    font-family: 'Courier New';
+    border-radius: 12px;
+}
+
+.GiftAllLogin {
+    background-color: pink;
+    width: 100%;
+    height: 630px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.titlePage {
+    width: 75%;
+}
+
+h1 {
+    font-size: 50px;
+}
+
+button {
+    background-color: lightcoral;
+    border: none;
+    text-align: center;
+    font-size: 25px;
+}
+
+.giftLogin {
+    margin-top: 5px;
+    padding: 7px;
+}
