+++ ClientApp/ClientApp/Components/Shared/_Header.razor
@@ -1,10 +1,10 @@
-﻿<div class="GiftingApp">
+﻿<div class="GiftAll">
     <nav id="mainNav">
-        <h1> Welcome To Gifting&Co </h1>
+        <h1> Welcome To GiftAll </h1>
         <ul>
             <li><a href="GiftHouseCreation.html">GiftHouseCreation</a></li>
         </ul>
     </nav>
 
 
-</div> <!-- end GiftingApp -->
+</div> <!-- end GiftAll -->
