+++ ClientApp/ClientApp/Components/_Imports.razor
@@ -9,3 +9,4 @@
 @using ClientApp.Components
 @using Microsoft.AspNetCore.Components.Authorization
 @using Microsoft.AspNetCore.Authorization
+@using ClientApp.Models;
